+++ ReprocessingDataBaseline/distribution/build_docker.bash
@@ -4,6 +4,8 @@ VERSION="1.0.0"
 
 CUR_DIR="$( cd "$(dirname "$0")" ; pwd -P )"
 
+shopt -s extglob 
+
 docker image inspect global_build:01.00.00 > /dev/null
 if [ $? -ne 0 ]; then
     echo "Building docker for build"
@@ -23,7 +25,7 @@ cd ${CUR_DIR}/docker_build
 rm -r source*
 
 mkdir source
-cp -r ../../../ReprocessingDataBaseline/* source/
+cp -r ../../../ReprocessingDataBaseline/!(distribution) source/
 
 docker run --rm -v ${CUR_DIR}/docker_build/source:/source global_build:01.00.00
 
@@ -32,7 +34,7 @@ cd ${CUR_DIR}
 
 cd docker_binary
 
-rm -r tmp_dir
+[ -d tmp_dir ] && rm -r tmp_dir
 
 mkdir tmp_dir
